Extract plugin-api-validation module
The protocol transaction validation contracts (TransactionValidatorService
and the TransactionValidationRule SPI) move to a new
besu-plugin-api-validation artifact, re-exported by besu-plugin-api so
existing consumers are unaffected. Their javadoc now states where the rule
is enforced and when it must be registered. TransactionValidatorServiceImpl
relocates to ethereum.mainnet.pluginadapter next to the machinery that runs
the rules.
