+++ ethereum/core/plugin-api-validation/build.gradle
@@ -1,0 +1,53 @@
+/*
+ * Copyright contributors to Besu.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+apply plugin: 'java-library'
+
+// override the path-derived name (besu-ethereum-core-plugin-api-validation) with the plugin API family naming
+def validationApiArtifactId = 'besu-plugin-api-validation'
+
+jar {
+  archiveBaseName = validationApiArtifactId
+  manifest {
+    attributes(
+      'Specification-Title': archiveBaseName,
+      'Specification-Version': project.version,
+      'Implementation-Title': archiveBaseName,
+      'Implementation-Version': calculateVersion(),
+      'Commit-Hash': getGitCommitDetails(40).hash,
+      'Automatic-Module-Name': 'org.hyperledger.besu.plugin.api.validation'
+      )
+  }
+}
+
+dependencies {
+  api project(':plugin-api:core')
+  // Transaction appears in this module's own signatures, so the dependency is declared
+  // here even though plugin-api:core also re-exports it
+  api project(':datatypes')
+}
+
+publishing {
+  publications {
+    mavenJava(MavenPublication) {
+      groupId = 'org.hyperledger.besu'
+      artifactId = validationApiArtifactId
+      pom {
+        name = 'Besu Plugin API Validation'
+        description = 'Protocol validation contracts of the Besu plugin API, letting plugins register additional validation rules; today these are transaction validation rules, applied wherever the node validates a transaction'
+      }
+    }
+  }
+}
