+++ plugin-api/build.gradle
@@ -42,6 +42,7 @@ dependencies {
   api project(':ethereum:core:plugin-api-worldstate-backend')
   api project(':ethereum:core:plugin-api-execution')
   api project(':ethereum:blockcreation:plugin-api')
+  api project(':ethereum:core:plugin-api-validation')
   api project(':datatypes')
   api 'org.apache.commons:commons-lang3'
   api 'io.consensys.tuweni:tuweni-bytes'
@@ -82,7 +83,8 @@ tasks.register('checkAPICompatibility', me.champeau.gradle.japicmp.JapicmpTask) 
     project(':ethereum:core:plugin-api-chain').tasks.named('jar'),
     project(':ethereum:core:plugin-api-worldstate-backend').tasks.named('jar'),
     project(':ethereum:core:plugin-api-execution').tasks.named('jar'),
-    project(':ethereum:blockcreation:plugin-api').tasks.named('jar'))
+    project(':ethereum:blockcreation:plugin-api').tasks.named('jar'),
+    project(':ethereum:core:plugin-api-validation').tasks.named('jar'))
   newClasspath.from(sourceSets.main.runtimeClasspath)
 
   packageIncludes = [
