+++ settings.gradle
@@ -46,6 +46,7 @@ include 'ethereum:blockcreation:plugin-api'
 include 'ethereum:core'
 include 'ethereum:core:plugin-api-chain'
 include 'ethereum:core:plugin-api-execution'
+include 'ethereum:core:plugin-api-validation'
 include 'ethereum:core:plugin-api-worldstate'
 include 'ethereum:core:plugin-api-worldstate-backend'
 include 'ethereum:eth'
